# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/invoices:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [billing]
      summary: Create a billing invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/invoices/{id}:
    get:
      operationId: getInvoice
      tags: [billing]
      summary: Get invoice details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [billing]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/invoices/{id}/pay:
    post:
      operationId: recordPayment
      tags: [billing]
      summary: Record payment for an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/charge-types:
    get:
      operationId: listChargeTypes
      tags: [billing]
      summary: List billing charge types (consultation, procedure, etc.)
      responses:
        "200":
          description: Charge types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChargeType"
    post:
      operationId: createChargeType
      tags: [billing]
      summary: Add a charge type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargeTypeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargeType"

  # ── CERTIFICATES ──────────────────────────────────────────────────────────
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List medical certificates
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Certificates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateListResponse"
    post:
      operationId: createCertificate
      tags: [certificates]
      summary: Generate a medical certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Certificate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/{id}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get certificate details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  # ── AUDIT LOGS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    InvoiceItem:
      type: object
      required: [description, quantity, unitPrice, total]
      properties:
        chargeTypeId:
          type: ["string", "null"]
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
    Invoice:
      type: object
      required: [id, invoiceNumber, patientId, items, subtotal, total, status, createdAt]
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        patientId:
          type: string
        patientName:
          type: string
        consultationId:
          type: ["string", "null"]
        doctorId:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        paymentMode:
          type: ["string", "null"]
          enum: [cash, card, upi, insurance, null]
        status:
          type: string
          enum: [draft, pending, paid, partial, cancelled, refunded]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    InvoiceInput:
      type: object
      required: [patientId, items]
      properties:
        patientId:
          type: string
        consultationId:
          type: string
        doctorId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        discount:
          type: number
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        discount:
          type: number
        notes:
          type: string
        status:
          type: string
    PaymentInput:
      type: object
      required: [amount, paymentMode]
      properties:
        amount:
          type: number
        paymentMode:
          type: string
          enum: [cash, card, upi, insurance]
        transactionReference:
          type: string
        notes:
          type: string
    InvoiceListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ChargeType:
      type: object
      required: [id, name, category, unitPrice]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [consultation, procedure, investigation, other]
        unitPrice:
          type: number
        taxPercent:
          type: number
        isActive:
          type: boolean
    ChargeTypeInput:
      type: object
      required: [name, category, unitPrice]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [consultation, procedure, investigation, other]
        unitPrice:
          type: number
        taxPercent:
          type: number

    # ── CERTIFICATE SCHEMAS ──────────────────────────────────────────────────
    Certificate:
      type: object
      required: [id, patientId, doctorId, type, issuedDate, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        patientName:
          type: string
        doctorId:
          type: string
        doctorName:
          type: string
        consultationId:
          type: ["string", "null"]
        type:
          type: string
          enum: [sick_leave, fitness, medical, procedure, vaccination]
        issuedDate:
          type: string
        fromDate:
          type: ["string", "null"]
        toDate:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        qrCode:
          type: ["string", "null"]
        createdAt:
          type: string
    CertificateInput:
      type: object
      required: [patientId, doctorId, type, issuedDate]
      properties:
        patientId:
          type: string
        doctorId:
          type: string
        consultationId:
          type: string
        type:
          type: string
          enum: [sick_leave, fitness, medical, procedure, vaccination]
        issuedDate:
          type: string
        fromDate:
          type: string
        toDate:
          type: string
        diagnosis:
          type: string
        content:
          type: string
    CertificateListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── AUDIT LOG SCHEMAS ────────────────────────────────────────────────────
...